refactor(core): centralize helper semantics
